Keyboard action keys directly calls i3blocks scripts
Keyboard buttons such as brightness and volume control were bound to
the action command (light/amixer) and to the update of the corresponding
block in i3blocks. They are now bound directly to the i3blocks script
with the appropriate BLOCK_BUTTON variable set, as if it was a mouse
event in the i3bar. This allows more DRY and maintainable code in the
future.