Remove ~/.dotfiles symlink
The install script installs the dotfiles from wherever it is launched.
There is no need to force create the ~/.dotfiles symlink.
With the old configuration, the directory was not overwritten, and thus
generated an error on each run. The directive has been removed, and
therefore no error is shown on a clean install.