Somehow fix i3blocks freeze on IP address retrieval
Couldn't reproduce: when getting internal or external IP address with
right or middle click on the INTERNET block, the whole i3blocks instance
freezes with no message to journald. It seems a program (printf ??)
writes something to stderr and it triggers i3blocks.
Redirecting stderr output to /dev/null solves the problem.