i3blocks: use more concise muted volume indicator
The volume i3blocks script, when in muted state, now displays it volume
but with a strikethrough style, rather than with the words "muted". This
allows the block to remain at the same width, regardless of its state.
Dirty hack warning: I replaced two of the separator spaces to include
them in individual scripts, in order to gain more space on the muted
volume script. This is rather ugly but, hey, aren't all my scripts
already ugly?