i3blocks music: refactor and show more music info in smaller font
Reduce code duplication and check if the part after the second " - "
contains useful info about the song, such as remix or live version.
Usually, text after the second " - " from Spotify isn't relevant because
it often consists of "Remastered" or "Original Mix". However, it
sometimes carry useful information, which I try to filter and display in
a smaller font size to gain space.